backend/requirements_renamed.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
logfire==0.15.1
pydantic==2.5.3
openai==1.12.0
python-multipart==0.0.9 